+++ database/migrations/2016_12_17_119816_create_categorizables_table.php
@@ -13,7 +13,7 @@ class CreateCategorizablesTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create(config('rinvex.categories.tables.categorizables'), function (Blueprint $table) {
             // Columns
@@ -33,7 +33,7 @@ class CreateCategorizablesTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists(config('rinvex.categories.tables.categorizables'));
     }