+++ src/Console/Commands/PublishCommand.php
@@ -6,21 +6,21 @@ namespace Rinvex\Categories\Console\Commands;
 
 use Illuminate\Console\Command;
 
-class MigrateCommand extends Command
+class PublishCommand extends Command
 {
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'rinvex:migrate:categories {--force : Force the operation to run when in production.}';
+    protected $signature = 'rinvex:publish:categories {--force : Overwrite any existing files.}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Migrate Rinvex Categories Tables.';
+    protected $description = 'Publish Rinvex Categories Resources.';
 
     /**
      * Execute the console command.
@@ -30,6 +30,6 @@ class MigrateCommand extends Command
     public function handle(): void
     {
         $this->warn($this->description);
-        $this->call('migrate', ['--step' => true, '--path' => 'vendor/rinvex/categories/database/migrations', '--force' => $this->option('force')]);
+        $this->call('vendor:publish', ['--tag' => 'rinvex-categories-config', '--force' => $this->option('force')]);
     }
 }